logs: Write the method in 3rd position, before the path
In recent commit 2285d2a4, we added the method to the download logs.
We've put it after the request path, as such: `method:<METHOD>`. So it
looked like that:

    2025/04/01 06:51:07.686160 REDIRECT 302 "/README" method:HEAD ip:1.2.3.4 mirror:mirror.foobar asn:1234 distance:42km countries:AB

After thinking a bit more, I'd like to adjust. Here's why.

The format `field:value` is useful when a field can be nil (`ip` can be
nil in case of error), or to distinguish between optional and mandatory
fields (for example: `mirror`, `asn`, `distance`, and `countries` are
optional, they are not present in case of 404).

However, we now log the method in all cases, and it's never nil, so
there's no good reason to log it with the `field:value` format.

So instead, let's log it as is, just before the request path.

After this commit, the line of log above looks like that:

    2025/04/01 06:51:07.686160 REDIRECT 302 HEAD "/README" ip:xxx.xxx.xxx.xxx mirror:mirror.foo.bar asn:1234 distance:42km countries:AB