cli: Fix incorrect comment
A URL without a scheme is NOT a relative URL.

Unfortunately I used this wrong term in a bunch of commit messages
before. Sorry for the confusion.

Fortunately I didn't use the term in any place of the code, except this
very comment, so fix it now.

For the avid reader: Wikipedia tells us that we can use the term
"Protocol-relative URLs":

> Protocol-relative links (PRL), also known as protocol-relative URLs
> (PRURL), are URLs that have no protocol specified. For example,
> `//example.com` will use the protocol of the current page, typically
> HTTP or HTTPS

However that's not what we use for HTTP+HTTPS mirrors, we use URLs that
have only a host and a path. I'm still looking for the right term for
that, if any...