config: Normalize URL for fallback mirrors as well
The function `NormalizeURL` is used to ensure that a mirror URL ends
with a trailing slash. It's applied when the mirror configuration is
written to the database (cf. `setMirror` in `rpc/rpc.go`).

Mirrorbits then relies on this fact when it constructs the redirect
URLs, in `pagerenderer.go`:

```
path := strings.TrimPrefix(results.FileInfo.Path, "/")
[...]
http.Redirect([...], results.MirrorList[0].AbsoluteURL+path, http.StatusFound)
```

However, it turns out that `NormalizeURL` is not applied to the fallback
URLs. Meaning that if your fallback is `http://mirror.org` and you
request the file `foobar`, and if ever the fallback mechanism kicks in,
mirrorbits will happily redirect you to `http://mirror.orgfoobar`...

Easy to fix! One less papercut!