scan: Log the number of files removed after the source was scanned
When a *mirror* is scanned, the number of files removed is logged:

```
2025/05/19 01:21:10.644 UTC [archive-4.kali.org] Requesting file list via rsync...
2025/05/19 01:22:42.670 UTC [archive-4.kali.org] Indexed 490848 files (490847 known), 3807 removed
```

However when the *source* is scanned, it's not logged:

```
2025/05/19 00:25:22.639 UTC [source] Scanning the filesystem...
2025/05/19 00:25:58.265 UTC [source] Indexing the files...
2025/05/19 00:26:02.360 UTC [source] Scanned 490848 files
```

The number of removed files can be useful information, and it's trivial
to add it to the log, so let's add it.

Additionally, this commit trades the term "Scanned" for "Indexed", to
align with the term that is used for the *mirror* scans (above).

After this commit, the logs now look like that:

```
2025/05/19 00:25:22.639 UTC [source] Scanning the filesystem...
2025/05/19 00:25:58.265 UTC [source] Indexing the files...
2025/05/19 00:26:02.360 UTC [source] Indexed 490848 files, 3807 removed
```