rsync: support rsync-ssl connections too
rsync provides a rsync-ssl script which tries to communicate over
tls, on port 874 instead of 873, if it finds openssl/gnutls/stunnel.

In theory we could add a separate cli option for this since there is
no official URL scheme for rsync+tls, but to keep things simple for
everyone just allow rsyncs:// in rsync URLs.

In case a URL is starting with rsyncs:// we will use rsync-ssl instead
of rsync, and rewrite the url back to rsync://.

For testing, here are two public servers that support TLS:

* rsync-ssl rsync://mirrors.dotsrc.org
* rsync-ssl rsync://repo.msys2.org

Using "--contimeout" with rsync-ssl leads to:
"The --contimeout option may only be used when connecting to an rsync daemon."

From my testing "--timeout" aborts though if openssl takes too long to connect,
suggesting that it covers both connection and i/o in this scenario.
Bump the --timeout to 60 so the upper bound is the same as without openssl.